big update
Added a getOptions in the class Ticket that returns 3 destination options;
Added a setDestination method that will compare the input String with the given options;
Added a print method that will print the ticket with destination, date and price.

Modified the TicketMachine class:
Added an arraylist of tickets;
Added a  printTicket method that will print a list of availble tickets from the list;
Added a method that allows the user to select a ticket from the list

+++ app02-ticketmachine/better-ticket-machine/Coin.java
@@ -13,14 +13,14 @@ public enum Coin
     P100 (100),
     P200 (200);
 
-    public final double value;
+    public final int value;
 
-    private Coin(double value)
+    private Coin(int value)
     {
         this.value = value;
     }
 
-    public double getValue()
+    public int getValue()
     {
         return value;
     }
